Move Attempts, Queries, and Clones into states module
