+++ proxy/coco/src/request.rs
@@ -1,4 +1,3 @@
-use std::ops::{Add, AddAssign};
 use std::{collections::HashMap, convert::TryFrom, marker::PhantomData, time::Duration};
 
 use either::Either;
@@ -14,75 +13,13 @@ pub use existential::SomeRequest;
 
 mod sealed;
 
-const MAX_QUERIES: Queries = Queries(1);
-const MAX_CLONES: Clones = Clones(1);
+const MAX_QUERIES: Queries = Queries::new(1);
+const MAX_CLONES: Clones = Clones::new(1);
 const PERIOD: Duration = Duration::from_secs(1); // Not for the whole request but for re-request
 
 pub fn exponential_backoff(attempt: usize, interval: Duration) -> Duration {
     let exp = u32::try_from(attempt).unwrap_or(u32::MAX);
     Duration::from_millis(u64::pow(2, exp)) + interval
-}
-
-#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct Queries(usize);
-
-impl Queries {
-    pub fn new(n: usize) -> Self {
-        Self(n)
-    }
-}
-
-impl Add<usize> for Queries {
-    type Output = Self;
-
-    fn add(self, other: usize) -> Self::Output {
-        Self(self.0 + other)
-    }
-}
-
-impl AddAssign<usize> for Queries {
-    fn add_assign(&mut self, other: usize) {
-        *self = Self(self.0 + other)
-    }
-}
-
-#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct Clones(usize);
-
-impl Clones {
-    pub fn new(n: usize) -> Self {
-        Self(n)
-    }
-}
-
-impl Add<usize> for Clones {
-    type Output = Self;
-
-    fn add(self, other: usize) -> Self::Output {
-        Self(self.0 + other)
-    }
-}
-
-impl AddAssign<usize> for Clones {
-    fn add_assign(&mut self, other: usize) {
-        *self = Self(self.0 + other)
-    }
-}
-
-#[derive(Clone, Copy, Debug, Hash, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Attempts {
-    queries: Queries, // how often we gossip
-    clones: Clones,   // how often we try to clone
-}
-
-impl Attempts {
-    pub fn new() -> Self {
-        Attempts {
-            queries: Queries(0),
-            clones: Clones(0),
-        }
-    }
 }
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
